# Optimize traverse