--- a/core/src/main/scala-2.13+/cats/instances/arraySeq.scala
+++ b/core/src/main/scala-2.13+/cats/instances/arraySeq.scala
@@ -234,7 +234,7 @@ private[cats] object ArraySeqInstances {
         G match {
           case x: StackSafeMonad[G] =>
             x.map(TraverseFilter.traverseFilterDirectly(fa.iterator)(f)(x))(
-              _.to(ArraySeq.untagged)
+              _.iterator.to(ArraySeq.untagged)
             )
           case _ =>
             fa.foldRight(Eval.now(G.pure(ArraySeq.untagged.empty[B]))) { case (x, xse) =>