--- a/core/src/main/scala-2.13+/cats/instances/arraySeq.scala
+++ b/core/src/main/scala-2.13+/cats/instances/arraySeq.scala
@@ -114,7 +114,7 @@ private[cats] object ArraySeqInstances {
         G match {
           case x: StackSafeMonad[G] => Traverse.traverse_Directly(fa)(f)(x)
           case _ =>
-            foldRight(fa, Always(G.pure(()))) { (a, acc) =>
+            foldRight(fa, Eval.now(G.unit)) { (a, acc) =>
               G.map2Eval(f(a), acc) { (_, _) =>
                 ()
               }