--- a/core/src/main/scala/cats/TraverseFilter.scala
+++ b/core/src/main/scala/cats/TraverseFilter.scala
@@ -21,7 +21,7 @@
 
 package cats
 
-import cats.data.State
+import cats.data.{Chain, State}
 import cats.kernel.compat.scalaVersionSpecific._
 
 import scala.collection.immutable.{IntMap, TreeSet}
@@ -206,8 +206,8 @@ object TraverseFilter {
 
   private[cats] def traverseFilterDirectly[G[_], A, B](
     fa: IterableOnce[A]
-  )(f: A => G[Option[B]])(implicit G: StackSafeMonad[G]): G[Vector[B]] = {
-    fa.iterator.foldLeft(G.pure(Vector.empty[B])) { case (bldrG, a) =>
+  )(f: A => G[Option[B]])(implicit G: StackSafeMonad[G]): G[Chain[B]] = {
+    fa.iterator.foldLeft(G.pure(Chain.empty[B])) { case (bldrG, a) =>
       G.map2(bldrG, f(a)) {
         case (acc, Some(b)) => acc :+ b
         case (acc, None)    => acc