--- a/core/src/main/scala/cats/data/Chain.scala
+++ b/core/src/main/scala/cats/data/Chain.scala
@@ -1258,7 +1258,7 @@ sealed abstract private[data] class ChainInstances extends ChainInstances1 {
         G match {
           case x: StackSafeMonad[G] => Traverse.traverse_Directly(fa.iterator)(f)(x)
           case _ =>
-            foldRight(fa, Always(G.pure(()))) { (a, acc) =>
+            foldRight(fa, Eval.now(G.unit)) { (a, acc) =>
               G.map2Eval(f(a), acc) { (_, _) =>
                 ()
               }