… bench/src/main/scala/cats/bench/TraverseBench.scala
… core/src/main/scala/cats/instances/map.scala
… core/src/main/scala/cats/instances/queue.scala
… core/src/main/scala/cats/instances/seq.scala
… core/src/main/scala/cats/instances/vector.scala
… tests/shared/src/test/scala-2.13+/cats/tests/ArraySeqSuite.scala
… tests/shared/src/test/scala/cats/tests/ChainSuite.scala
… tests/shared/src/test/scala/cats/tests/ListSuite.scala
… tests/shared/src/test/scala/cats/tests/MapSuite.scala
… tests/shared/src/test/scala/cats/tests/QueueSuite.scala
… tests/shared/src/test/scala/cats/tests/SeqSuite.scala
… tests/shared/src/test/scala/cats/tests/VectorSuite.scala