This addresses the `StackSafeMonad` part of https://github.com/typelevel/cats/issues/4408. I thought that there were enough changes in this to merit a separate PR for the `Defer` part.

Open questions (will edit these out of the description later):
- The `TraverseFilter` laws hard-code `Option`. How do we test that the new branching implementation is lawful?

The following benchmarks were run on an AWS t3.xlarge:

## Baseline (b08196e8526f6025d009ac0c8efc0a0e0ebcb1c4)

|Benchmark             |              (length) |  Mode | Cnt   |   Score  |   Error |  Units |
| --- | --- | --- | --- | --- | --- | --- |
|TraverseBench.filterList        |       10000 | thrpt |  20 | 3247.486| ± 270.817 | ops/s|
|TraverseBench.filterVector       |      10000  |thrpt  | 20 | 3351.585| ± 255.883 | ops/s|
|TraverseBench.mapList          |        10000 | thrpt  | 20 | 3198.905 |± 235.865 | ops/s|
|TraverseBench.mapVector   |             10000 | thrpt  | 20 | 3906.339| ± 225.508 | ops/s|
|TraverseBench.traverseChain     |       10000 | thrpt  | 20  | 556.706| ±  31.814 | ops/s|
|TraverseBench.traverseChainError    |   10000 | thrpt  | 20 | 1278.954 |±  74.242 | ops/s |
|TraverseBench.traverseFilterChain  |    10000 | thrpt  | 20 |  536.841 |±  38.696  |ops/s |
|TraverseBench.traverseFilterList   |    10000 | thrpt  | 20 |  532.103 |±  30.473 | ops/s |
|TraverseBench.traverseFilterVector  |   10000 | thrpt |  20  | 590.312 |±  29.534 | ops/s |
|TraverseBench.traverseList      |       10000 | thrpt |  20  | 537.023| ±  33.287 | ops/s |
|TraverseBench.traverseListError    |    10000 | thrpt |  20 | 1674.930| ±  48.771 | ops/s |
|TraverseBench.traverseVector      |     10000 | thrpt |  20 |  581.844 |±  32.074  |ops/s |
|TraverseBench.traverseVectorError   |   10000 | thrpt |  20 | 1778.334 |± 135.859 | ops/s |
|TraverseBench.traverse_Chain   |        10000 | thrpt  | 20|   490.822| ±  14.145 | ops/s |
|TraverseBench.traverse_List    |        10000 | thrpt  | 20 |  474.568 |±  28.995 | ops/s |
|TraverseBench.traverse_Vector   |       10000  |thrpt  | 20 |  622.924| ±  38.002  |ops/s |


## Chain (2d5f4d7917804e921488f5f2e263e8349a6ceb42)

|Benchmark             |              (length) |  Mode | Cnt   |   Score  |   Error |  Units |
| --- | --- | --- | --- | --- | --- | --- |
|TraverseBench.filterList       |        10000 | thrpt |  20  |3170.974| ± 146.628 | ops/s |
|TraverseBench.filterVector    |         10000 | thrpt  | 20 | 3217.881 |± 181.087  | ops/s |
| TraverseBench.mapList       |           10000|  thrpt  | 20 | 3443.969| ± 104.117  | ops/s |
| TraverseBench.mapVector       |         10000 | thrpt |  20 | 3827.401 | ± 148.074 | ops/s |
|TraverseBench.traverseChain    |        10000 | thrpt |  20 |  784.845| ±  25.990  | ops/s |
|TraverseBench.traverseChainError  |     10000 | thrpt  | 20 | 1302.782| ±  62.342  | ops/s |
|TraverseBench.traverseFilterChain   |   10000 | thrpt |  20 |  748.986| ±  62.151 |  ops/s |
|TraverseBench.traverseFilterList   |    10000 | thrpt  | 20  | 669.817| ±  41.798 | ops/s |
|TraverseBench.traverseFilterVector  |   10000 | thrpt |  20  | 618.644 |±  28.558 | ops/s |
|TraverseBench.traverseList     |        10000 | thrpt |  20 |  583.413 | ±  27.079 | ops/s |
| TraverseBench.traverseListError  |      10000 | thrpt  | 20 | 1253.510| ±  76.328 | ops/s |
|TraverseBench.traverseVector    |       10000 | thrpt |  20  | 581.975 | ±  27.619  | ops/s |
|TraverseBench.traverseVectorError  |    10000 | thrpt |  20  |1302.362 |±  66.015 | ops/s |
|TraverseBench.traverse_Chain    |       10000 | thrpt  | 20 | 3598.033 |± 108.420 | ops/s |
|TraverseBench.traverse_List    |        10000 | thrpt |  20 | 4048.287 | ± 158.090 | ops/s |
|TraverseBench.traverse_Vector   |       10000 | thrpt  | 20 | 4094.024 | ± 101.115 | ops/s |

## Vector (702ab8b9d1ebad207bb7c97eb671da333e667ea0)

|Benchmark             |              (length) |  Mode | Cnt   |   Score  |   Error |  Units |
| --- | --- | --- | --- | --- | --- | --- |
|TraverseBench.filterList         |      10000 |  thrpt |  20 | 3305.299 |± 233.483 | ops/s |
|TraverseBench.filterVector       |      10000 | thrpt  | 20 | 3253.531 |± 133.993 | ops/s |
|TraverseBench.mapList     |             10000 | thrpt  | 20 | 3308.961| ± 138.697|  ops/s |
|TraverseBench.mapVector      |          10000 | thrpt |  20 | 3911.187 |± 142.582 | ops/s|
|TraverseBench.traverseChain       |     10000 | thrpt  | 20  | 594.535 |±  29.777|  ops/s|
|TraverseBench.traverseChainError    |   10000 | thrpt  | 20  |1084.548| ±  75.686 | ops/s |
|TraverseBench.traverseFilterChain  |    10000 | thrpt  | 20 |  642.895| ±  55.621 | ops/s |
|TraverseBench.traverseFilterList    |   10000 | thrpt |  20  | 682.965 |±  41.234 | ops/s |
|TraverseBench.traverseFilterVector  |   10000 | thrpt |  20 |  696.235| ±  59.945  |ops/s |
|TraverseBench.traverseList     |        10000 | thrpt  | 20  | 582.275 |±  35.877 | ops/s |
|TraverseBench.traverseListError   |     10000 | thrpt |  20 | 1114.959| ±  52.641 | ops/s |
|TraverseBench.traverseVector     |      10000 | thrpt  | 20 |  638.235| ±  42.533 | ops/s |
|TraverseBench.traverseVectorError   |   10000 | thrpt |  20  |1098.038 |±  62.946 | ops/s |
|TraverseBench.traverse_Chain     |      10000 | thrpt |  20 | 3480.086 |± 178.654 | ops/s |
|TraverseBench.traverse_List     |       10000 | thrpt |  20  |3950.650 |± 171.441 | ops/s |
|TraverseBench.traverse_Vector     |     10000|  thrpt  | 20 | 3695.495| ± 182.401 | ops/s |