@johnynek 2023-08-23
Can you post the results of the benchmarks? Apologies if I missed them.