@armanbilge 2023-11-08
> Can you give an explanation as to why we shouldn't worry about mutability in such a case?

We can't have concurrency: we are sequencing the effects with `flatMap`, which by law must be sequential.