@armanbilge 2023-11-10
Oscar is right: using mutable data structures to implement traverse will not work, but perhaps surprisingly the issue is not concurrency or laziness. I added a test in 2210d8271792a14c20b5b25b7b54b82f464779f6 that fails when tested against the builder-based implementation in a1c9ff5d02f744cc6789627877c98059e6b4b3a5. There is some cheating there to pretend that `List` is a stacksafe monad just so that we invoke the `traverseDirectly` codepath.

As it turns out, the problem is actually here:
https://github.com/typelevel/cats/blob/a1c9ff5d02f744cc6789627877c98059e6b4b3a5/core/src/main/scala/cats/Traverse.scala#L299-L302

For `List`-like monads, you may invoke that map function multiple times for each element of the `List`. If the `acc`umulator is shared mutable state then this breaks down very quickly, which is exactly what that test demonstrates.

Suffice to say: my bad 😅