@TimWSpence 2023-11-14
Thanks @johnynek and @armanbilge I'm convinced that we should use immutable data structures 😆 I've run the benchmarks for the immutable vector variant. I'll do the same for immutable list and chain - I don't have a good intuition which would be best?

Also thanks for the other comments - I'll address those too :)