@TimWSpence 2023-12-13
OK I've run the benchmarks for the `Chain` implementation and the `traverse_` performance dropped significantly. But I didn't touch the implementation of `traverse_`... I've running these on an M2 Max. I've heard of benchmarks ending up running on its efficiency cores - could that be what happened?

Also realized that I didn't update `TraverseFilter` to use `Chain` 😭