@armanbilge 2023-12-13
> I've running these on an M2 Max

<img width="733" alt="Screen Shot 2023-12-13 at 8 32 01 AM" src="https://github.com/typelevel/cats/assets/3119428/bc41c4c1-1fad-4178-8b0b-e8bb6590f40b">

> Final warning — do not under any circumstances do this on Apple hardware.

https://www.youtube.com/watch?v=n5u7DgFwLGE&t=1541s