@TimWSpence 2023-12-14
> > I've running these on an M2 Max
> 
> <img alt="Screen Shot 2023-12-13 at 8 32 01 AM" width="733" src="https://private-user-images.githubusercontent.com/3119428/290261204-bc41c4c1-1fad-4178-8b0b-e8bb6590f40b.png?jwt=eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpc3MiOiJnaXRodWIuY29tIiwiYXVkIjoicmF3LmdpdGh1YnVzZXJjb250ZW50LmNvbSIsImtleSI6ImtleTEiLCJleHAiOjE3MDI1Njk4OTgsIm5iZiI6MTcwMjU2OTU5OCwicGF0aCI6Ii8zMTE5NDI4LzI5MDI2MTIwNC1iYzQxYzRjMS0xZmFkLTQxNzgtOGIwYi1lOGJiNjU5MGY0MGIucG5nP1gtQW16LUFsZ29yaXRobT1BV1M0LUhNQUMtU0hBMjU2JlgtQW16LUNyZWRlbnRpYWw9QUtJQUlXTkpZQVg0Q1NWRUg1M0ElMkYyMDIzMTIxNCUyRnVzLWVhc3QtMSUyRnMzJTJGYXdzNF9yZXF1ZXN0JlgtQW16LURhdGU9MjAyMzEyMTRUMTU1OTU4WiZYLUFtei1FeHBpcmVzPTMwMCZYLUFtei1TaWduYXR1cmU9N2E3YWI3NDA5ZDQ4NmZiMzQ0ZjM0OTI2MGJjZmEwNGVjNGU0MDgyNzMxOGI3NTJlNDIzMWMxZGYyNTUwZGQzNyZYLUFtei1TaWduZWRIZWFkZXJzPWhvc3QmYWN0b3JfaWQ9MCZrZXlfaWQ9MCZyZXBvX2lkPTAifQ.BirAz0JuVTALMB8g4vJFQBursSW1xWrctTVX7Og05Qs">
> 
> > Final warning — do not under any circumstances do this on Apple hardware.
> 
> https://www.youtube.com/watch?v=n5u7DgFwLGE&t=1541s

😆 Sadly it's all I have. Any suggestions?