@TimWSpence 2023-12-18
OK I've now got access to a relatively chonky EC2 instance so I'll try to get the full set of benchmarks done ASAP