@TimWSpence 2023-12-22
OK there are the benchmarks for baseline, chain and vector implementations. The `traverse_` improvements are looking pretty sweet but the `traverse` improvements are currently very small (especially when we have to convert between data types). I'll try to investigate in the new year but if anyone has ideas in the meantime then I'd love to hear them!