@johnynek 2023-12-24
Am I wrong that the Chain version seems to beat the Vector version?