@TimWSpence 2024-03-04
@armanbilge @johnynek @valencik sorry for letting this languish for so long 😭 I've now reverted the last change which was just to get some baseline benchmarks and merged `main` in again.

Just wanted to gauge people's thoughts on the best direction from here. Assuming that my benchmarks are reliable (if anyone else wants to run them as well, that would be fantastic!), the results are a bit disappointing for `traverse` (baseline, chain and vector all look relatively similar. The new versions are slightly better, especially for `Chain` collections with the `Chain` implementation and similarly for `Vector`) but pretty great for `traverse_`. Where do you think we should go from here?