@TimWSpence 2024-03-04
> I listed a couple of minor nits, but I think any performance improvement is worth merging, especially around traverse since it is such a key function.

Thanks! Do you think it's worth making separate `Chain` and `Vector` implementations respectively or just implement them all with `Chain` as that seems to be the best overall?