@johnynek 2024-03-04
I would go where the benchmarks suggest if we have confidence in them.